chore: duplicates multiclient and rpc_config in chain/evm
