+++ .changeset/cool-sheep-grow.md
@@ -1,0 +1,5 @@
+---
+"chainlink-deployments-framework": patch
+---
+
+chore: duplicates multiclient and rpc_config in chain/evm
