+++ .changeset/short-pigs-cheer.md
@@ -1,0 +1,5 @@
+---
+"chainlink-deployments-framework": minor
+---
+
+Remove migration file from the scaffold. Pipelines is the preferred way to run Changesets.
