+++ chain/evm/multiclient_test.go
@@ -1,4 +1,4 @@
-package deployment
+package evm
 
 import (
 	"context"
