+++ chain/evm/provider/rpc_provider.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/smartcontractkit/chainlink-deployments-framework/chain"
 	"github.com/smartcontractkit/chainlink-deployments-framework/chain/evm"
-	"github.com/smartcontractkit/chainlink-deployments-framework/deployment"
 )
 
 // RPCChainProviderConfig holds the configuration to initialize the RPCChainProvider.
@@ -21,7 +20,7 @@ type RPCChainProviderConfig struct {
 	// key from a private key, or TransactorFromKMS to create a deployer key from a KMS key.
 	DeployerTransactorGen SignerGenerator
 	// Required: At least one RPC must be provided to connect to the EVM node.
-	RPCs []deployment.RPC
+	RPCs []evm.RPC
 	// Required: ConfirmFunctor is a type that generates a confirmation function for transactions.
 	// Use ConfirmFuncGeth to use the Geth client for transaction confirmation, or
 	// ConfirmFuncSeth to use the Seth client for transaction confirmation with richer debugging.
@@ -32,7 +31,7 @@ type RPCChainProviderConfig struct {
 	// RPCChainProvider. These options are applied to the MultiClient instance created by the
 	// RPCChainProvider. You can use this to set up custom HTTP clients, timeouts, or other
 	// configurations for the RPC connections.
-	ClientOpts []func(client *deployment.MultiClient)
+	ClientOpts []func(client *evm.MultiClient)
 	// Optional: A generator for the additional user transactors. If not provided, no user
 	// transactors will be generated.
 	UsersTransactorGen []SignerGenerator
@@ -124,9 +123,7 @@ func (p *RPCChainProvider) Initialize(ctx context.Context) (chain.BlockChain, er
 	}
 
 	// Setup the client.
-	//
-	// TODO: Move the multiclient into this package
-	client, err := deployment.NewMultiClient(p.config.Logger, deployment.RPCConfig{
+	client, err := evm.NewMultiClient(p.config.Logger, evm.RPCConfig{
 		ChainSelector: p.selector,
 		RPCs:          p.config.RPCs,
 	}, p.config.ClientOpts...)
