+++ chain/evm/provider/rpc_provider_test.go
@@ -10,17 +10,16 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/smartcontractkit/chainlink-deployments-framework/chain/evm"
-	"github.com/smartcontractkit/chainlink-deployments-framework/deployment"
 )
 
 func Test_RPCChainProviderConfig_validate(t *testing.T) {
 	t.Parallel()
 
-	rpc := deployment.RPC{
+	rpc := evm.RPC{
 		Name:               "Test",
 		HTTPURL:            "http://localhost:8545",
 		WSURL:              "ws://localhost:8546",
-		PreferredURLScheme: deployment.URLSchemePreferenceHTTP,
+		PreferredURLScheme: evm.URLSchemePreferenceHTTP,
 	}
 
 	confirmFuncGeth := ConfirmFuncGeth(10 * time.Millisecond)
@@ -34,14 +33,14 @@ func Test_RPCChainProviderConfig_validate(t *testing.T) {
 			name: "valid config",
 			config: RPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        confirmFuncGeth,
 			},
 		},
 		{
 			name: "missing deployer transactor generator",
 			config: RPCChainProviderConfig{
-				RPCs:           []deployment.RPC{rpc},
+				RPCs:           []evm.RPC{rpc},
 				ConfirmFunctor: confirmFuncGeth,
 			},
 			wantErr: "deployer transactor generator is required",
@@ -50,7 +49,7 @@ func Test_RPCChainProviderConfig_validate(t *testing.T) {
 			name: "missing confirm functor",
 			config: RPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 			},
 			wantErr: "confirm functor is required",
 		},
@@ -90,10 +89,10 @@ func Test_RPCChainProvider_Initialize(t *testing.T) {
 	mockSrv := newFakeRPCServer(t)
 
 	// Define a general RPC configuration for use
-	rpc := deployment.RPC{
+	rpc := evm.RPC{
 		Name:               "Test",
 		HTTPURL:            mockSrv.URL,
-		PreferredURLScheme: deployment.URLSchemePreferenceHTTP,
+		PreferredURLScheme: evm.URLSchemePreferenceHTTP,
 	}
 
 	gethConfirmFunc := ConfirmFuncGeth(1 * time.Second)
@@ -110,7 +109,7 @@ func Test_RPCChainProvider_Initialize(t *testing.T) {
 			giveSelector: chainSelector,
 			giveConfig: RPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				UsersTransactorGen: []SignerGenerator{
 					TransactorRandom(),
 				},
@@ -122,7 +121,7 @@ func Test_RPCChainProvider_Initialize(t *testing.T) {
 			giveSelector: chainSelector,
 			giveConfig: RPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				Logger:                logger.Test(t),
 				ConfirmFunctor:        gethConfirmFunc,
 			},
@@ -132,7 +131,7 @@ func Test_RPCChainProvider_Initialize(t *testing.T) {
 			giveSelector: chainSelector,
 			giveConfig: RPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor: ConfirmFuncSeth(
 					rpc.HTTPURL, 10*time.Millisecond, []string{}, configPath,
 				),
@@ -154,7 +153,7 @@ func Test_RPCChainProvider_Initialize(t *testing.T) {
 			giveSelector: 1, // Invalid selector
 			giveConfig: RPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        gethConfirmFunc,
 			},
 			wantErr: "failed to get chain ID from selector",
@@ -164,7 +163,7 @@ func Test_RPCChainProvider_Initialize(t *testing.T) {
 			giveSelector: chainSelector,
 			giveConfig: RPCChainProviderConfig{
 				DeployerTransactorGen: &alwaysFailingTransactorGenerator{},
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        gethConfirmFunc,
 			},
 			wantErr: "failed to generate deployer key",
@@ -174,7 +173,7 @@ func Test_RPCChainProvider_Initialize(t *testing.T) {
 			giveSelector: chainSelector,
 			giveConfig: RPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				UsersTransactorGen: []SignerGenerator{
 					&alwaysFailingTransactorGenerator{}, // This will always fail
 				},
@@ -187,7 +186,7 @@ func Test_RPCChainProvider_Initialize(t *testing.T) {
 			giveSelector: chainSelector,
 			giveConfig: RPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{{}},
+				RPCs:                  []evm.RPC{{}},
 				ConfirmFunctor:        gethConfirmFunc,
 			},
 			wantErr: "failed to create multi-client",
@@ -197,7 +196,7 @@ func Test_RPCChainProvider_Initialize(t *testing.T) {
 			giveSelector: chainSelector,
 			giveConfig: RPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor: ConfirmFuncSeth(
 					rpc.HTTPURL, 10*time.Millisecond, []string{}, "nonexistent.toml",
 				), // Invalid path
@@ -209,7 +208,7 @@ func Test_RPCChainProvider_Initialize(t *testing.T) {
 			giveSelector: chainSelector,
 			giveConfig: RPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        &alwaysFailingConfirmFunctor{},
 			},
 			wantErr: "failed to generate confirm function",
