+++ chain/evm/provider/zksync_rpc_provider.go
@@ -13,7 +13,6 @@ import (
 
 	"github.com/smartcontractkit/chainlink-deployments-framework/chain"
 	"github.com/smartcontractkit/chainlink-deployments-framework/chain/evm"
-	"github.com/smartcontractkit/chainlink-deployments-framework/deployment"
 )
 
 // ZkSyncRPCChainProviderConfig holds the configuration to initialize the RPCChainProvider. While
@@ -29,7 +28,7 @@ type ZkSyncRPCChainProviderConfig struct {
 	// TransactorFromKMS, you should use ZkSyncSignerFromKMS.
 	ZkSyncSignerGen ZkSyncSignerGenerator
 	// Required: At least one RPC must be provided to connect to the EVM node.
-	RPCs []deployment.RPC
+	RPCs []evm.RPC
 	// Required: ConfirmFunctor is a type that generates a confirmation function for transactions.
 	// Use ConfirmFuncGeth to use the Geth client for transaction confirmation, or
 	// ConfirmFuncSeth to use the Seth client for transaction confirmation with richer debugging.
@@ -40,7 +39,7 @@ type ZkSyncRPCChainProviderConfig struct {
 	// RPCChainProvider. These options are applied to the MultiClient instance created by the
 	// RPCChainProvider. You can use this to set up custom HTTP clients, timeouts, or other
 	// configurations for the RPC connections.
-	ClientOpts []func(client *deployment.MultiClient)
+	ClientOpts []func(client *evm.MultiClient)
 	// Optional: Logger is the logger to use for the RPCChainProvider. If not provided, a default
 	// logger will be used.
 	Logger logger.Logger
@@ -124,7 +123,7 @@ func (p *ZkSyncRPCChainProvider) Initialize(ctx context.Context) (chain.BlockCha
 	}
 
 	// Setup the client.
-	client, err := deployment.NewMultiClient(p.config.Logger, deployment.RPCConfig{
+	client, err := evm.NewMultiClient(p.config.Logger, evm.RPCConfig{
 		ChainSelector: p.selector,
 		RPCs:          p.config.RPCs,
 	}, p.config.ClientOpts...)
