+++ chain/evm/provider/zksync_rpc_provider_test.go
@@ -10,17 +10,16 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/smartcontractkit/chainlink-deployments-framework/chain/evm"
-	"github.com/smartcontractkit/chainlink-deployments-framework/deployment"
 )
 
 func Test_ZkSyncRPCChainProviderConfig_validate(t *testing.T) {
 	t.Parallel()
 
-	rpc := deployment.RPC{
+	rpc := evm.RPC{
 		Name:               "Test",
 		HTTPURL:            "http://localhost:8545",
 		WSURL:              "ws://localhost:8546",
-		PreferredURLScheme: deployment.URLSchemePreferenceHTTP,
+		PreferredURLScheme: evm.URLSchemePreferenceHTTP,
 	}
 
 	confirmFuncGeth := ConfirmFuncGeth(10 * time.Millisecond)
@@ -35,7 +34,7 @@ func Test_ZkSyncRPCChainProviderConfig_validate(t *testing.T) {
 			config: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
 				ZkSyncSignerGen:       ZKSyncSignerRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        confirmFuncGeth,
 			},
 		},
@@ -43,7 +42,7 @@ func Test_ZkSyncRPCChainProviderConfig_validate(t *testing.T) {
 			name: "missing deployer transactor generator",
 			config: ZkSyncRPCChainProviderConfig{
 				ZkSyncSignerGen: ZKSyncSignerRandom(),
-				RPCs:            []deployment.RPC{rpc},
+				RPCs:            []evm.RPC{rpc},
 				ConfirmFunctor:  confirmFuncGeth,
 			},
 			wantErr: "deployer transactor generator is required",
@@ -52,7 +51,7 @@ func Test_ZkSyncRPCChainProviderConfig_validate(t *testing.T) {
 			name: "missing signer generator",
 			config: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        confirmFuncGeth,
 			},
 			wantErr: "signer generator is required",
@@ -62,7 +61,7 @@ func Test_ZkSyncRPCChainProviderConfig_validate(t *testing.T) {
 			config: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
 				ZkSyncSignerGen:       ZKSyncSignerRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 			},
 			wantErr: "confirm functor is required",
 		},
@@ -103,10 +102,10 @@ func Test_ZkSyncRPCChainProvider_Initialize(t *testing.T) {
 	mockSrv := newFakeRPCServer(t)
 
 	// Define a general RPC configuration for use
-	rpc := deployment.RPC{
+	rpc := evm.RPC{
 		Name:               "Test",
 		HTTPURL:            mockSrv.URL,
-		PreferredURLScheme: deployment.URLSchemePreferenceHTTP,
+		PreferredURLScheme: evm.URLSchemePreferenceHTTP,
 	}
 
 	gethConfirmFunc := ConfirmFuncGeth(1 * time.Second)
@@ -124,7 +123,7 @@ func Test_ZkSyncRPCChainProvider_Initialize(t *testing.T) {
 			giveConfig: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
 				ZkSyncSignerGen:       ZKSyncSignerRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        gethConfirmFunc,
 			},
 		},
@@ -134,7 +133,7 @@ func Test_ZkSyncRPCChainProvider_Initialize(t *testing.T) {
 			giveConfig: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
 				ZkSyncSignerGen:       ZKSyncSignerRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				Logger:                logger.Test(t),
 				ConfirmFunctor:        gethConfirmFunc,
 			},
@@ -145,7 +144,7 @@ func Test_ZkSyncRPCChainProvider_Initialize(t *testing.T) {
 			giveConfig: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
 				ZkSyncSignerGen:       ZKSyncSignerRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor: ConfirmFuncSeth(
 					rpc.HTTPURL, 10*time.Millisecond, []string{}, configPath,
 				),
@@ -168,7 +167,7 @@ func Test_ZkSyncRPCChainProvider_Initialize(t *testing.T) {
 			giveConfig: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
 				ZkSyncSignerGen:       ZKSyncSignerRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        gethConfirmFunc,
 			},
 			wantErr: "failed to get chain ID from selector",
@@ -179,7 +178,7 @@ func Test_ZkSyncRPCChainProvider_Initialize(t *testing.T) {
 			giveConfig: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: &alwaysFailingTransactorGenerator{},
 				ZkSyncSignerGen:       ZKSyncSignerRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        gethConfirmFunc,
 			},
 			wantErr: "failed to generate deployer key",
@@ -190,7 +189,7 @@ func Test_ZkSyncRPCChainProvider_Initialize(t *testing.T) {
 			giveConfig: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
 				ZkSyncSignerGen:       ZKSyncSignerRandom(),
-				RPCs:                  []deployment.RPC{{}},
+				RPCs:                  []evm.RPC{{}},
 				ConfirmFunctor:        gethConfirmFunc,
 			},
 			wantErr: "failed to create multi-client",
@@ -201,7 +200,7 @@ func Test_ZkSyncRPCChainProvider_Initialize(t *testing.T) {
 			giveConfig: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
 				ZkSyncSignerGen:       ZKSyncSignerRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor: ConfirmFuncSeth(
 					rpc.HTTPURL, 10*time.Millisecond, []string{}, "nonexistent.toml",
 				), // Invalid path
@@ -214,7 +213,7 @@ func Test_ZkSyncRPCChainProvider_Initialize(t *testing.T) {
 			giveConfig: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
 				ZkSyncSignerGen:       ZKSyncSignerRandom(),
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        &alwaysFailingConfirmFunctor{},
 			},
 			wantErr: "failed to generate confirm function",
@@ -225,7 +224,7 @@ func Test_ZkSyncRPCChainProvider_Initialize(t *testing.T) {
 			giveConfig: ZkSyncRPCChainProviderConfig{
 				DeployerTransactorGen: TransactorRandom(),
 				ZkSyncSignerGen:       &alwaysFailingZKSyncSignerGenerator{},
-				RPCs:                  []deployment.RPC{rpc},
+				RPCs:                  []evm.RPC{rpc},
 				ConfirmFunctor:        gethConfirmFunc,
 			},
 			wantErr: "failed to generate zkSync signer",
