+++ chain/evm/rpc_config_test.go
@@ -1,4 +1,4 @@
-package deployment
+package evm
 
 import (
 	"testing"
