+++ engine/cld/scaffold/scaffold.go
@@ -131,11 +131,8 @@ func ScaffoldEnvDir(envdir cldf_domain.EnvDir) error {
 		}),
 		fileFSNode(cldf_domain.NodesFileName, withTemplate("nodes.json.tmpl")),
 		fileFSNode(cldf_domain.ViewStateFileName, withTemplate("view_state.json.tmpl")),
-		fileFSNode(cldf_domain.MigrationsFileName, withTemplate("migrations.go.tmpl")),
-		fileFSNode("migrations_test.go", withTemplate("migrations_test.go.tmpl")),
-		fileFSNode(cldf_domain.MigrationsArchiveFileName, withTemplate("migrations_archive.go.tmpl")),
-		fileFSNode(cldf_domain.DurablePipelinesFileName, withTemplate("durable_pipelines.go.tmpl")),
-		fileFSNode("durable_pipelines_test.go", withTemplate("durable_pipelines_test.go.tmpl")),
+		fileFSNode(cldf_domain.PipelinesFileName, withTemplate("pipelines.go.tmpl")),
+		fileFSNode("pipelines_test.go", withTemplate("pipelines_test.go.tmpl")),
 		dirFSNode(cldf_domain.ArtifactsDirName, []*fsnode{gitKeepFSNode()}),
 		dirFSNode(cldf_domain.ProposalsDirName, []*fsnode{gitKeepFSNode()}),
 		dirFSNode(cldf_domain.ArchivedProposalsDirName, []*fsnode{gitKeepFSNode()}),
