--- engine/cld/scaffold/templates/migrations.go.tmpl
@@ -1,36 +1,0 @@
-package {{.package}}
-
-import (
-	cldf_changeset "github.com/smartcontractkit/chainlink-deployments-framework/engine/cld/changeset"
-)
-
-var _ cldf_changeset.RegistryProvider = (*MigrationsRegistryProvider)(nil)
-
-// MigrationsRegistryProvider wraps the cldf_changeset.BaseRegistryProvider and
-// provides an Init function to register migrations with the underlying
-// registry.
-type MigrationsRegistryProvider struct {
-	*cldf_changeset.BaseRegistryProvider
-}
-
-// NewMigrationsRegistryProvider creates a new MigrationsRegistryProvider. The
-// BaseRegistryProvider is initialized with a default registry.
-func NewMigrationsRegistryProvider() *MigrationsRegistryProvider {
-	return &MigrationsRegistryProvider{
-		BaseRegistryProvider: cldf_changeset.NewBaseRegistryProvider(),
-	}
-}
-
-// Init is used to register migrations with the registry
-//
-// Add your migrations to the registry here.
-func (p *MigrationsRegistryProvider) Init() error {
-	// Archives old migrations defined in migrations_archive.go
-	p.Archive()
-
-	// Uncomment this line to start adding new migrations to the registry
-	// lint:ignore Intentionally unused for future migrations
-	// registry := p.Registry()
-
-	return nil
-}
