--- engine/cld/scaffold/templates/migrations_test.go.tmpl
@@ -1,1 +1,0 @@
-package {{.package}}
