+++ engine/cld/scaffold/templates/pipelines.go.tmpl
@@ -1,0 +1,35 @@
+package {{.package}}
+
+import (
+	cldf_changeset "github.com/smartcontractkit/chainlink-deployments-framework/engine/cld/changeset"
+)
+
+var _ cldf_changeset.RegistryProvider = (*PipelinesRegistryProvider)(nil)
+
+// PipelinesRegistryProvider wraps the cldf_changeset.BaseRegistryProvider and
+// provides an Init function to register Pipelines with the underlying registry.
+type PipelinesRegistryProvider struct {
+	*cldf_changeset.BaseRegistryProvider
+}
+
+// NewPipelinesRegistryProvider creates a new PipelinesRegistryProvider.
+// The BaseRegistryProvider is initialized with a default registry.
+func NewPipelinesRegistryProvider() *PipelinesRegistryProvider {
+	return &PipelinesRegistryProvider{
+		BaseRegistryProvider: cldf_changeset.NewBaseRegistryProvider(),
+	}
+}
+
+// Archive is a noop for the DurablePipelines.
+func (*PipelinesRegistryProvider) Archive() {}
+
+// Init is used to register Durable Pipelines with the registry
+//
+// Add your Durable Pipeline migration to the registry here.
+func (p *PipelinesRegistryProvider) Init() error {
+	// Uncomment this line to start adding new Pipelines to the registry
+	// lint:ignore Intentionally unused for future migrations
+	// registry := p.Registry()
+
+	return nil
+}
