+++ engine/cld/scaffold/templates/pipelines_test.go.tmpl
